Major patch browser update
Patch browser updated in preparation for PD 0.52.0
Fully functional patch browser with use of new [file] object requires PD 0.52.0

*Known Bug* 
When closing the patch mouse_mouse will appear in the console. 
This is remints of a object that was deleted. A help ticket has been submitted for this bug.
This does not effect the overall functionality of this patch